Fix compatibility issue with Math.floorMod

+++ test/framework/src/main/java/org/elasticsearch/test/ESTestCase.java
@@ -1448,7 +1448,7 @@ public abstract class ESTestCase extends LuceneTestCase {
         } else {
             // we adjust the gradle worker id with mod so as to not go over the ephemoral port ranges, but gradle continually
             // increases this value, so the mod can eventually become zero, thus we shift on both sides by 1
-            final long workerId = Long.valueOf(workerIdStr);
+            final int workerId = Integer.valueOf(workerIdStr);
             assert workerId >= 1 : "Non positive gradle worker id: " + workerIdStr;
             startAt = (int) Math.floorMod(workerId - 1, 223L) + 1;
         }
